requirements.txt
requests>=2.31.0
tqdm>=4.66.0
PyYAML>=6.0.1
python-dateutil>=2.9.0.post0
mwparserfromhell>=0.6.6
